+++ lib/ruby_jard/layouts.rb
@@ -1,0 +1,30 @@
+# frozen_string_literal: true
+
+module RubyJard
+  ##
+  # Layouts registry.
+  module Layouts
+    class << self
+      def layout_registry
+        @layout_registry ||= {}
+      end
+
+      def add_layout(name, layout_class)
+        unless layout_class.is_a?(RubyJard::Templates::LayoutTemplate)
+          raise RubyJard::Error, "#{layout_class} must be a #{RubyJard::Templates::LayoutTemplate}"
+        end
+
+        layout_registry[name] = layout_class
+      end
+
+      def [](name)
+        layout_registry[name.to_s.strip]
+      end
+      alias_method :get, :[]
+
+      def each(&block)
+        @layout_registry.each(&block)
+      end
+    end
+  end
+end