+++ lib/ruby_jard/screens/source_screen.rb
@@ -6,9 +6,9 @@ module RubyJard
     # Display source code of current stopping line and surrounding lines
     class SourceScreen < RubyJard::Screen
       def title
-        return 'Source' if RubyJard.current_session.frame.nil?
+        return 'Source' if @session.current_frame.nil?
 
-        decorated_path = path_decorator(current_file, current_line)
+        decorated_path = path_decorator(@session.current_frame.frame_file, @session.current_frame.frame_line)
         if decorated_path.gem?
           ['Source', "#{decorated_path.gem} - #{decorated_path.path}:#{decorated_path.lineno}"]
         else
@@ -17,7 +17,7 @@ module RubyJard
       end
 
       def build
-        return if RubyJard.current_session.frame.nil?
+        return if @session.current_frame.nil?
 
         # TODO: screen now supports window.
         codes = source_decorator.codes
@@ -45,7 +45,7 @@ module RubyJard
         lineno = source_lineno(index)
         RubyJard::Span.new(
           margin_right: 1,
-          content: current_line == lineno ? '➠' : ' ',
+          content: @session.current_frame.frame_line == lineno ? '➠' : ' ',
           styles: :source_line_mark
         )
       end
@@ -54,39 +54,25 @@ module RubyJard
         lineno = source_lineno(index).to_s.rjust(source_decorator.window_end.to_s.length)
         RubyJard::Span.new(
           content: lineno,
-          styles: current_line == lineno ? :source_line_mark : :source_lineno
+          styles: @session.current_frame.frame_line == lineno ? :source_line_mark : :source_lineno
         )
       end
 
       def loc_spans(loc)
-        spans, _tokens = loc_decorator.decorate(loc, current_file)
+        spans, _tokens = loc_decorator.decorate(loc, @session.current_frame.frame_file)
         spans
       end
 
       private
-
-      def current_binding
-        RubyJard.current_session.frame._binding
-      end
-
-      def current_frame_scope
-        RubyJard.current_session.backtrace[RubyJard.current_session.frame.pos][1]
-      end
-
-      def current_file
-        RubyJard.current_session.frame.file
-      end
-
-      def current_line
-        RubyJard.current_session.frame.line
-      end
 
       def path_decorator(path, lineno)
         @path_decorator ||= RubyJard::Decorators::PathDecorator.new(path, lineno)
       end
 
       def source_decorator
-        @source_decorator ||= RubyJard::Decorators::SourceDecorator.new(current_file, current_line, @layout.height)
+        @source_decorator ||= RubyJard::Decorators::SourceDecorator.new(
+          @session.current_frame.frame_file, @session.current_frame.frame_line, @layout.height
+        )
       end
 
       def loc_decorator